+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 2.5.9
     - Symmetrize TSNE affinities matrix, fixes optimization
     - Add numerical stability to SoftPlus activation function
+    - Fix Extra Tree leaf node entropy
 
 - 2.5.8
     - Optimize squaring operations